pyproject.toml
[project]
name = "rudimentary"
version = "0.1.0"
description = "Synthetic dataset generator for PAS drum rudiments"
requires-python = ">=3.10"
dependencies = [
    "mido>=1.3.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pyfluidsynth>=1.3.0",
    "pedalboard>=0.7.0",
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
    "pyarrow>=14.0.0",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "matplotlib>=3.7.0",
]
cloud = [
    "ray>=2.8.0",
]
hub = [
    "huggingface_hub>=0.20.0",
    "datasets>=2.16.0",
]
ml = [
    # Core ML
    "torch>=2.0",
    "torchaudio>=2.0",
    "transformers>=4.35",
    # Training & logging
    "tensorboard>=2.14",
    # Analysis
    "scikit-learn>=1.3",
    "seaborn>=0.12",
    # Optional but recommended
    # "wandb>=0.16",  # Uncomment for weights & biases logging
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dataset_gen"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
# Allow unused variables in specific patterns
ignore = [
    "E741",  # Ambiguous variable names (I, l, O) - acceptable in mathematical code
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F841"]  # Unused variables in tests (common for checking side effects)
# Unused variables in algorithm implementations (computed for documentation/future use)
"dataset_gen/audio_aug/chain.py" = ["F841"]
"dataset_gen/audio_aug/degradation.py" = ["F841"]
"dataset_gen/labels/groove.py" = ["F841"]
"dataset_gen/midi_gen/generator.py" = ["F841"]

[tool.black]
line-length = 100
target-version = ["py310"]
